requirements.txt
questionary>=1.10.0
jinja2>=3.1.2
pyyaml>=6.0.1
pathlib>=1.0.1 